# Add benchmark for `RndState` and add non-null tracers in benchmark