--- a/lib/core/cardano-wallet-core.cabal
+++ b/lib/core/cardano-wallet-core.cabal
@@ -336,6 +336,7 @@ benchmark db
     , persistent-sqlite
     , random
     , temporary
+    , text
     , time
     , transformers
   type: